Update due to the creation of a new class

+++ Calculations.cs
@@ -15,7 +15,7 @@ namespace Awose
             if (current > max) return max;
             return current;
         }
-        public static bool IsInRadius(int CurX, int CurY, AwoseAgent obj, float radius)
+        public static bool IsInRadius(float CurX, float CurY, AwoseAgent obj, float radius)
         {
             return (Math.Sqrt(Math.Pow(CurX - obj.X, 2) + Math.Pow(CurY - obj.Y, 2)) <= radius);
         }
